requirements.txt
transformers
torch
torchvision
pillow
opencv-python
timm